requirements.txt
numpy
matplotlib
jupyter
keras
tensorflow-rocm
Pillow
scikit-learn